Revert to auto-loading driver jars
We have many clients that expect simply including the jdbc-* gems is
sufficient to ensure the actual JDBC driver is on the classpath.
While we want to move to requiring an intentional load_driver, or
loading the driver when it's actually needed at connection time,
we can't break compatibility in the 1.2.x series.

For now, if you want to explicitly tell the jdbc-* gems to not auto-load
the driver jars, set the Java system property arjdbc.skip.autoload to true.